tests for king graphs complete
