+++ .gitignore
@@ -1,7 +1,8 @@
 debug.cpp
 a
 experiment_runner
-graphs/
+build/
+/graphs/
 cses.cpp
 test/
 *.venv
