+++ tests/graphs/grid5.gr
@@ -1,0 +1,105 @@
+p sp 25 104
+a 0 1 1
+a 0 6 1.41421
+a 1 0 1
+a 1 2 1
+a 1 7 1.41421
+a 1 5 1.41421
+a 2 1 1
+a 2 3 1
+a 2 8 1.41421
+a 2 6 1.41421
+a 3 2 1
+a 3 4 1
+a 3 9 1.41421
+a 3 7 1.41421
+a 4 3 1
+a 4 8 1.41421
+a 5 1 1.41421
+a 5 6 1
+a 5 11 1.41421
+a 6 0 1.41421
+a 6 2 1.41421
+a 6 5 1
+a 6 7 1
+a 6 12 1.41421
+a 6 10 1.41421
+a 7 1 1.41421
+a 7 3 1.41421
+a 7 6 1
+a 7 8 1
+a 7 13 1.41421
+a 7 11 1.41421
+a 8 2 1.41421
+a 8 4 1.41421
+a 8 7 1
+a 8 9 1
+a 8 14 1.41421
+a 8 12 1.41421
+a 9 3 1.41421
+a 9 8 1
+a 9 13 1.41421
+a 10 6 1.41421
+a 10 11 1
+a 10 16 1.41421
+a 11 5 1.41421
+a 11 7 1.41421
+a 11 10 1
+a 11 12 1
+a 11 17 1.41421
+a 11 15 1.41421
+a 12 6 1.41421
+a 12 8 1.41421
+a 12 11 1
+a 12 13 1
+a 12 18 1.41421
+a 12 16 1.41421
+a 13 7 1.41421
+a 13 9 1.41421
+a 13 12 1
+a 13 14 1
+a 13 19 1.41421
+a 13 17 1.41421
+a 14 8 1.41421
+a 14 13 1
+a 14 18 1.41421
+a 15 11 1.41421
+a 15 16 1
+a 15 21 1.41421
+a 16 10 1.41421
+a 16 12 1.41421
+a 16 15 1
+a 16 17 1
+a 16 22 1.41421
+a 16 20 1.41421
+a 17 11 1.41421
+a 17 13 1.41421
+a 17 16 1
+a 17 18 1
+a 17 23 1.41421
+a 17 21 1.41421
+a 18 12 1.41421
+a 18 14 1.41421
+a 18 17 1
+a 18 19 1
+a 18 24 1.41421
+a 18 22 1.41421
+a 19 13 1.41421
+a 19 18 1
+a 19 23 1.41421
+a 20 16 1.41421
+a 20 21 1
+a 21 15 1.41421
+a 21 17 1.41421
+a 21 20 1
+a 21 22 1
+a 22 16 1.41421
+a 22 18 1.41421
+a 22 21 1
+a 22 23 1
+a 23 17 1.41421
+a 23 19 1.41421
+a 23 22 1
+a 23 24 1
+a 24 18 1.41421
+a 24 23 1
